Backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
idna==3.10
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
stripe==11.1.1
typing_extensions==4.12.2
urllib3==2.2.3
